+++ Reinvest/Registration/src/Adapter/Database/RegistrationSchema.ts
@@ -1,6 +1,7 @@
 import { JSONObjectOf } from 'HKEKTypes/Generics';
 import { Insertable } from 'kysely';
 import { NorthCapitalEntityType, NorthCapitalSynchronizationRecordType } from 'Registration/Adapter/NorthCapital/NorthCapitalSynchronizationRecord';
+import { PlaidResult } from 'Registration/Domain/Model/BankAccount';
 import { MappedRecordType } from 'Registration/Domain/Model/Mapping/MappedRecord';
 import { MappedRecordStatus, MappedType } from 'Registration/Domain/Model/Mapping/MappedType';
 import { DocumentSyncState, NorthCapitalObjectType } from 'Registration/Domain/VendorModel/NorthCapital/NorthCapitalTypes';
@@ -57,15 +58,15 @@ export interface VertaloSynchronizationTable {
 
 export interface RegistrationBankAccountTable {
   accountId: string;
-  bankAccountId: string | null;
+  bankAccountId: string;
   bankAccountNumber: string | null;
   bankAccountType: string | null;
-  createdDate: Date;
   northCapitalId: string;
-  plaidJson: JSONObjectOf<any>;
+  plaidJson: JSONObjectOf<PlaidResult>;
   plaidUrl: string | null;
   profileId: string;
-  state: 'ACTIVE' | 'INACTIVE';
+  state: 'ACTIVE' | 'INACTIVE' | 'IN_PROGRESS';
+  createdDate?: Date;
 }
 
 export type InsertableMappingRegistry = Insertable<MappingRegistryTable>;